# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: VibeNode API
  version: 1.0.0
  description: |
    VibeNode is a GUI application for managing Claude Code sessions, projects,
    kanban boards, composition workflows, and more. This spec documents all
    REST endpoints served by the Flask web server on port 5050.

    **Base URL:** `http://localhost:5050`

    VibeNode also uses SocketIO for real-time push events. Those are documented
    separately in `socketio-events.md`.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5050
    description: Local development server
...
paths:
...
  /api/set-project:
    post:
      tags: [Projects]
      summary: Switch the active project
      operationId: setProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project]
              properties:
                project:
                  type: string
                  description: Encoded project identifier
      responses:
        "200":
          description: Project switched
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  project:
                    type: string
                  display:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/rename-project:
    post:
      tags: [Projects]
      summary: Set a custom display name for a project
      operationId: renameProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [encoded]
              properties:
                encoded:
                  type: string
                name:
                  type: string
                  description: Custom name (empty to clear)
      responses:
        "200":
          description: Renamed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/delete-project:
    post:
      tags: [Projects]
      summary: Delete a project and all its sessions
      operationId: deleteProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [encoded]
              properties:
                encoded:
                  type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/add-project:
    post:
      tags: [Projects]
      summary: Add a project via browse, path, or create new
      operationId: addProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
                  enum: [browse, path, create]
                  default: browse
                path:
                  type: string
                  description: Required for mode=path
                name:
                  type: string
                  description: Required for mode=create
      responses:
        "200":
          description: Project added
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  encoded:
                    type: string
                  path:
                    type: string
                  cancelled:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/project-chat:
    post:
      tags: [Projects]
      summary: AI-assisted project finder
      operationId: projectChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
      responses:
        "200":
          description: Chat response with suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                  suggestions:
                    type: array
                    items:
                      type: string
                  matches:
                    type: array
                    items:
                      type: object
  /api/new-session:
    post:
      tags: [Projects]
      summary: Launch a new Claude SDK session (or resume existing)
      operationId: newSession
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                resume_id:
                  type: string
                  description: If provided, resume this session instead of creating new
                prompt:
                  type: string
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  new_id:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  # ==========================================================================
  # Live (live_api.py)
  # ==========================================================================
...
  /api/compose/projects:
    post:
      tags: [Compose]
      summary: Create a new composition project
      operationId: createComposeProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                parent_project:
                  type: string
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  project:
                    type: object
        "400":
          $ref: "#/components/responses/BadRequest"
    get:
      tags: [Compose]
      summary: List all composition projects
      operationId: listComposeProjects
      parameters:
        - name: project
          in: query
          schema:
            type: string
          description: Filter by parent project
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  projects:
                    type: array
                    items:
                      type: object
  /api/compose/projects/{project_id}:
    get:
      tags: [Compose]
      summary: Get project detail with sections and context
      operationId: getComposeProject
      parameters:
        - $ref: "#/components/parameters/composeProjectId"
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Compose]
      summary: Update project settings
      operationId: updateComposeProject
      parameters:
        - $ref: "#/components/parameters/composeProjectId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                shared_prompts_enabled:
                  type: boolean
                root_session_id:
                  type: string
                position:
                  type: integer
                pinned:
                  type: boolean
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  project:
                    type: object
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Compose]
      summary: Delete a composition project and its folder
      operationId: deleteComposeProject
      parameters:
        - $ref: "#/components/parameters/composeProjectId"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  deleted:
                    type: boolean
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    composeProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
      description: Compose project UUID
...
  schemas:
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        ok:
          type: boolean
          example: false
...
  responses:
    BadRequest:
      description: Bad request — missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"